+++ arch/arm/src/imxrt/imxrt_tickless.c
@@ -501,7 +501,7 @@ int up_timer_gettime(struct timespec *ts)
 int up_alarm_start(const struct timespec *ts)
 {
   size_t offset = 1;
-  uint64_t tm = ((uint64_t)ts->tv_sec * NSEC_PER_SEC + ts->tv_nsec) /
+  uint64_t tm = (ts->tv_sec * NSEC_PER_SEC + ts->tv_nsec) /
                 NSEC_PER_TICK;
   irqstate_t flags;
   uint32_t regval;
