+++ arch/risc-v/src/bl602/bl602_oneshot_lowerhalf.c
@@ -124,7 +124,7 @@ static inline uint64_t bl602_get_nsec(void)
 
   up_timer_gettime(&ts);
 
-  return (uint64_t)ts.tv_nsec + (uint64_t)ts.tv_sec * NSEC_PER_SEC;
+  return ts.tv_nsec + ts.tv_sec * NSEC_PER_SEC;
 }
 
 /****************************************************************************
